Restructured ilitoppingmaker with imports in __init__.py
